Refactor impact and report processings into mixins
